Consolidate InviteClientContext and InviteServerContext
